+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - CUE background mode controlled by ConnectPRO with CueBackground config element on ConnectProTheme
 
 ## [9.0.0] - 2019-08-15
 ### Added